+++ modules/frontend/src/main/scala/components/ResultFilter.scala
@@ -14,43 +14,13 @@
  * limitations under the License.
  */
 
-package rallyeye
+package components
 
 import typings.flowbite.mod.initFlowbite
 
 import com.raquo.laminar.api.L._
+import rallyeye.Router
 import rallyeye.shared.RallyData
-
-object Components:
-  def header(rallySignal: Signal[RallyData], filterSignal: Signal[String]) =
-    navTag(
-      cls := "bg-white border-gray-200 px-4 lg:px-6 py-2.5 shadow-lg row-start-1 col-start-1",
-      div(
-        cls := "flex flex-wrap justify-between items-center mx-auto max-w-screen-xl",
-        a(
-          href := "https://rallyeye.2m.lt",
-          cls := "flex items-center",
-          img(src := "/rallyeye.svg", cls := "mr-3 h-6 sm:h-9", alt := "RallyEye logo"),
-          span(cls := "self-center text-xl font-semibold whitespace-nowrap", "RallyEye")
-        ),
-        child <-- rallySignal.map { r =>
-          div(
-            a(href := r.link, target := "_blank", r.name),
-            p(cls := "text-xs text-gray-400", "Data retrieved at ", r.retrievedAt.toString)
-          )
-        },
-        children <-- rallySignal.combineWith(filterSignal).map(ResultFilter.render),
-        div(
-          cls := "flex items-center lg:order-2",
-          a(
-            href := "https://github.com/2m/rallyeye/",
-            target := "_blank",
-            cls := "text-gray-800 hover:bg-gray-50 focus:ring-4 focus:ring-gray-300 font-medium rounded-lg text-sm px-4 lg:px-5 py-2 lg:py-2.5 mr-2 focus:outline-none",
-            "GitHub"
-          )
-        )
-      )
-    )
 
 object ResultFilter:
   val AllResults = "All Results"
